IClarification, ICommentary

+++ ContestModel/src/org/icpc/tools/contest/model/IClarification.java
@@ -45,7 +45,7 @@ public interface IClarification extends IContestObject {
 	 *
 	 * @return the contest time
 	 */
-	int getContestTime();
+	long getContestTime();
 
 	/**
 	 * Returns the wall clock time, in ms since the epoch.
